+++ packages/frontend/src/components/atoms/index.ts
@@ -1,8 +1,7 @@
-import Input from '../elements/Input';
 import Loader from './Loader';
 import LogOut from './LogOut';
 import UserInput from './UserInput';
 import UserSubmit from './UserSubmit';
 import MiniProduct from './MiniProduct';
 
-export { Input, Loader, LogOut, UserInput, UserSubmit, MiniProduct };
+export { Loader, LogOut, UserInput, UserSubmit, MiniProduct };
