+++ packages/frontend/src/components/elements/index.ts
@@ -1,3 +1,4 @@
 import Nav from './Nav';
+import Input from './Input';
 
-export { Nav };
+export { Nav, Input };
